Create assignment5.lisp
This is my assignment 5 lisp code from my previous semester. If it works perfectly, just add the macro functions needed for this assignment 5. Also, please check the error-checking part, I don't think I did that part for the assignments. Note: I put starpiece as my pieces to collect, but it's up to you guys for the pieces.